+++ src/main/java/com/mgh/backend/cashier/port/CashierSaleProduct.java
@@ -6,6 +6,6 @@ public record CashierSaleProduct(
         String barcode,
         String name,
         BigDecimal price,
-        int remainingStock
+        double remainingStock
 ) {
 }
